+++ README.md
@@ -1,1 +1,25 @@
 # Elice-Tree
+
+## elice-tree-be
+
+### How to Setup Requirements & Python Environments
+
+```sh
+### Create a virtual environment
+python3 -m venv venv
+
+### Activate the virtual environment
+source venv/bin/activate
+
+### Run the server
+python launcher.py
+
+### Install required dependencies from requirements.txt
+cd elice-tree-be
+pip install -r requirements.txt
+
+### (Optional) Upgrade pip to the latest version
+pip install --upgrade pip
+```
+
+## elice-tree-fe
