+++ elice-tree-be/app/main.py
@@ -1,5 +1,4 @@
 from fastapi import FastAPI
-import uvicorn
 from app.routers import auth, mission
 
 
@@ -21,8 +20,3 @@ FastAPI.openapi = openapi
 app = FastAPI()
 app.include_router(auth.router, tags=["auth"])
 app.include_router(mission.router, tags=["mission"])
-
-
-if __name__ == "__main__":
-    uvicorn.run(app, host="127.0.0.1", port=8080)
-# uvicorn app.main:app --reload --host 127.0.0.1 --port 8080
