+++ elice-tree-be/app/mock/data.py
@@ -5,5 +5,7 @@ mock_missions = [
         "id": 2,
         "content": "탄소배출을 위해 우리가 할 수 있는 행동",
         "isCompleted": False,
+        # "carbon": 1,
+        # "tree": 1,
     },
 ]
