+++ elice-tree-be/app/routers/auth.py
@@ -1,11 +1,16 @@
 from fastapi import APIRouter, HTTPException
 # from fastapi.security import APIKeyHeader
-from app.utils.auth import generate_access_token, verify_user_from_token
-from app.utils.function import make_base_response
 from app.mock.data import mock_user
 from app.schemas.base import BaseResponse
-from app.schemas.auth import AuthRequest, AuthResponse
-from app.schemas.example import (
+from app.schemas.auth import (
+    AuthRequest,
+    AuthResponse
+)
+from app.utils.function import (
+    generate_access_token,
+    make_base_response
+)
+from app.mock.example import (
     example_regist,
     example_login,
     example_400,
@@ -24,11 +29,11 @@ router = APIRouter()
 async def regist(request: AuthRequest):
     try:
         if request.id in mock_user:
-            return make_base_response(400, "User already exists", {})
+            return make_base_response(400, "User already exists")
         mock_user[request.id] = request.password
-        return make_base_response(200, "User registered successfully", {})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail="Internal Server Error") from e
+        return make_base_response(200, "User registered successfully")
+    except Exception:
+        return make_base_response(500, "Internal Server Error")
 
 
 @router.post("/api/auth/login", response_model=BaseResponse, responses={
@@ -41,17 +46,18 @@ async def login(request: AuthRequest):
         if request.id in mock_user and mock_user[request.id] == request.password:
             access_token = generate_access_token(data={"sub": request.id})
             return make_base_response(200, "Login successful", AuthResponse(accessToken=access_token))
-        return make_base_response(400, "Invalid credentials", {})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail="Internal Server Error") from e
+        return make_base_response(400, "Invalid credentials")
+    except Exception:
+        return make_base_response(500, "Internal Server Error")
 
 
+# TODO: /api/auth/logout
 # @router.post("/api/auth/logout", response_model=BaseResponse, responses={
 #     200: {"description": "User logged out successfully", "content": {"application/json": {"example": example_logout}}},
 #     500: {"description": "Internal Server Error", "content": {"application/json": {"example": example_500}}}
 # })
 # async def logout(token: str = Depends(APIKeyHeader(name="Authorization"))):
 #     try:
-#         return make_base_response(200, "Logout successfully", {})
+#         return make_base_response(200, "Logout successfully")
 #     except Exception as e:
-#         raise HTTPException(status_code=500, detail="Internal Server Error") from e
+#         return make_base_response(500, "Internal Server Error")
