+++ elice-tree-be/launcher.py
@@ -1,0 +1,5 @@
+import uvicorn
+
+if __name__ == "__main__":
+    uvicorn.run("app.main:app", host="127.0.0.1", port=8080, reload=True)
+# python launcher.py
